Reworked some things in components and objects. Fixed some errors which i somehow overlooked. Started to implement Space's, also added a rudimentary object allocator which is just a vector that acts fancy

+++ Dev/StaticSheepGame/SheepEngine/Handle.cpp
@@ -13,8 +13,8 @@ namespace Framework
   const Handle Handle::null;
 
   Handle::Handle( )
-    : m_index( -1 )
-    , m_counter( -1 )
+    : m_index( static_cast<size_t>(-1) )
+    , m_counter( static_cast<size_t>(-1) )
   {
   }
